registration/requirements.txt
asgiref==3.5.0
Django==3.2.12
django-dotenv==1.4.2
django-environ==0.8.1
django-filter==21.1
django-sendgrid-v5==1.2.0
Pillow==9.0.1
python-dotenv==0.20.0
python-http-client==3.3.7
pytz==2021.3
sendgrid==6.9.7
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
typing-extensions==4.1.1
